+Added Heterogeneous Euclidean-Overlap Metric (HEOM) +Added Heterogeneous Value Difference Metric (HVDM) *k Fold Cross Validation now splits in a stratified way *Fixed bug in ranges for Euclidean metric *Fixed bug in Utils when an arrf file has empty lines after the @data descriptor *Added function in Utils used to split a dataset into many vectors where each vector objects belong to the same class

+++ Machine Learning/Euclidean.h
@@ -13,5 +13,5 @@ public:
     void Normalize(const std::vector<Instance> &train);
 
 protected:
-    std::vector<float> ranges;
+    std::vector<double> ranges;
 };
